Unigap_project6/cloud-function-trigger/requirements.txt
google-cloud-bigquery
google-cloud-storage

